refactor: reorg API'
