+++ cognite_toolkit/__init__.py
@@ -1,0 +1,3 @@
+from ._api import CogniteToolkit, data_classes
+
+__all__ = ["CogniteToolkit", "data_classes"]
