+++ cognite_toolkit/_api/__init__.py
@@ -1,0 +1,4 @@
+from . import data_classes
+from .api_client import CogniteToolkit
+
+__all__ = ["CogniteToolkit", "data_classes"]
